+++ src/main/java/org/apache/rocketmq/dashboard/controller/OpsController.java
@@ -19,7 +19,7 @@ package org.apache.rocketmq.dashboard.controller;
 import com.google.common.base.Preconditions;
 import jakarta.annotation.Resource;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.rocketmq.dashboard.permisssion.Permission;
+import org.apache.rocketmq.dashboard.permission.Permission;
 import org.apache.rocketmq.dashboard.service.OpsService;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
