+++ src/main/java/org/apache/rocketmq/dashboard/controller/ProducerController.java
@@ -19,7 +19,7 @@ package org.apache.rocketmq.dashboard.controller;
 
 import jakarta.annotation.Resource;
 import org.apache.rocketmq.dashboard.model.ConnectionInfo;
-import org.apache.rocketmq.dashboard.permisssion.Permission;
+import org.apache.rocketmq.dashboard.permission.Permission;
 import org.apache.rocketmq.dashboard.service.ProducerService;
 import org.apache.rocketmq.remoting.protocol.body.ProducerConnection;
 import org.springframework.stereotype.Controller;
