+++ src/main/java/org/apache/rocketmq/dashboard/controller/ProxyController.java
@@ -17,7 +17,7 @@
 package org.apache.rocketmq.dashboard.controller;
 
 import jakarta.annotation.Resource;
-import org.apache.rocketmq.dashboard.permisssion.Permission;
+import org.apache.rocketmq.dashboard.permission.Permission;
 import org.apache.rocketmq.dashboard.service.ProxyService;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
